requirements.txt
asgiref==3.11.0
pytz==2025.2
sqlparse==0.5.5
Django==5.1.1
djangorestframework==3.15.2
PyJWT==2.10.1
djangorestframework_simplejwt==5.4.0
djoser==2.3.1